Clean up parameter system: remove dead code, use variant for defaults
Major improvements to the parameter/UI system:

1. Remove dead callback code from Parameter class
   - Removed unused add_change_callback/notify_callbacks mechanism
   - Actual callbacks use Configurable::on_parameter_changed()

2. Use variant for ControlLayout::default_val
   - Changed from int64_t to std::variant<int, bool, uint32_t, const char*, std::string>
   - Allows type-safe defaults without casts
   - String defaults now work: .default_val = "text"
   - Color defaults work: .default_val = 0xFF000000

3. Move effect enums from ui.h to effect_base.h
   - BlendMode, DrawStyle, AudioChannel, VerticalPosition, RenderMode
   - These are render logic, not UI definitions

4. Use ParameterType::ENUM for dropdowns/radio groups
   - Previously created INT parameters, now properly uses ENUM
   - Stores option labels for self-documenting parameters

5. Fix DynamicMovementEffect redundant parameter creation
   - Removed manual add_parameter calls that duplicated UI layout
   - Default pixel script now in UI layout

+++ libs/avs_lib/effects/clear_effect.cpp
@@ -77,7 +77,7 @@ const PluginInfo ClearEffect::effect_info {
                 .text = "Color",
                 .type = ControlType::COLOR_BUTTON,
                 .x = 0, .y = 15, .w = 137, .h = 13,
-                .default_val = static_cast<int>(0xFF000000)
+                .default_val = 0xFF000000
             },
             {
                 .id = "only_first",
